Removed ItemClickRecyclerView. Unecessary and now media gallery clicks handled in its adapter
